refactor(cyber_webapp): rename ontology_v2 -> ontology, drop PRIORS shim
  * Rename `cyber_webapp.ontology_v2` to `cyber_webapp.ontology` --
    the parallel-path marker is no longer needed (the only ontology
    module the pack ships is this one). Update every consumer in the
    pack + tests + ontology docstring.
  * Drop the back-compat `PRIORS = MappingProxyType(...)` shim from
    `cyber_webapp.priors` (and its now-unused `MappingProxyType`
    import). Phase 2 callers all migrated to `default_prior()` /
    `_CYBER_GENERATION_CONFIG`; nothing imports PRIORS anymore.
  * Tidy two stale doc references to long-deleted core modules.
